requirements.txt
telethon==1.34.0
pyyaml==6.0.1 